feat: Image controller & service 작성
