+++ friends_server/src/main/kotlin/com/friends/image/controller/ImageController.kt
@@ -1,0 +1,24 @@
+package com.friends.image.controller
+
+import com.friends.image.service.ImageCommandService
+import org.springframework.http.ResponseEntity
+import org.springframework.security.core.annotation.AuthenticationPrincipal
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestPart
+import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.multipart.MultipartFile
+
+@RestController
+@RequestMapping("/api/user/Image")
+class ImageController(
+    private val imageCommandService: ImageCommandService,
+) : ImageControllerSpec {
+    @PostMapping(consumes = ["multipart/form-data"])
+    override fun uploadImage(
+        @AuthenticationPrincipal
+        memberId: Long,
+        @RequestPart
+        image: MultipartFile,
+    ): ResponseEntity<String> = ResponseEntity.ok(imageCommandService.uploadImage(memberId, image))
+}
